Introduced a new feature - task streams
press Ctrl+S to open up a menu where notes can be added/removed to
various lists, which can be traversed for easy grouping/planning and
fast access. I'm not sure if it will work or not, and I anticipate
bugs. Luckily I am the only one using this software...